정적 팩토리 메소드로 변경
